Treat "permission denied" as a benign error.
os.ErrPermission was not treated a non-retryable error. This led to
directory listing not working if the mounting user did not have access
to one of the filesystems subdirectories when listing, as a global lock
would be held by the unsuccessful retry.

By passing through permission denied correctly, this behavior is resolved.